Move test only functions to testutil
