+++ beacon-chain/core/altair/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "epoch_spec.go",
         "fork_transition.go",
         "reward.go",
-        "state.go",
         "sync_committee.go",
         "transition.go",
         "validator.go",
@@ -30,7 +29,6 @@ go_library(
         "//beacon-chain/core/validators:go_default_library",
         "//beacon-chain/p2p/types:go_default_library",
         "//beacon-chain/state/interface:go_default_library",
-        "//beacon-chain/state/v1:go_default_library",
         "//beacon-chain/state/v2:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
@@ -44,7 +42,6 @@ go_library(
         "//shared/params:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
-        "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@io_opencensus_go//trace:go_default_library",
     ],
 )
@@ -60,8 +57,6 @@ go_test(
         "epoch_spec_test.go",
         "fork_transition_test.go",
         "reward_test.go",
-        "state_fuzz_test.go",
-        "state_test.go",
         "sync_committee_test.go",
         "transition_test.go",
         "validator_test.go",
@@ -70,7 +65,6 @@ go_test(
     deps = [
         "//beacon-chain/core/epoch/precompute:go_default_library",
         "//beacon-chain/core/helpers:go_default_library",
-        "//beacon-chain/core/state:go_default_library",
         "//beacon-chain/p2p/types:go_default_library",
         "//beacon-chain/state/interface:go_default_library",
         "//beacon-chain/state/v2:go_default_library",
@@ -80,7 +74,6 @@ go_test(
         "//shared/attestationutil:go_default_library",
         "//shared/bls:go_default_library",
         "//shared/bytesutil:go_default_library",
-        "//shared/hashutil:go_default_library",
         "//shared/interfaces:go_default_library",
         "//shared/params:go_default_library",
         "//shared/testutil:go_default_library",
