--- beacon-chain/core/altair/state.go
@@ -1,202 +1,0 @@
-package altair
-
-import (
-	"context"
-
-	"github.com/pkg/errors"
-	"github.com/prysmaticlabs/go-bitfield"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
-	iface "github.com/prysmaticlabs/prysm/beacon-chain/state/interface"
-	"github.com/prysmaticlabs/prysm/beacon-chain/state/v1"
-	stateAltair "github.com/prysmaticlabs/prysm/beacon-chain/state/v2"
-	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
-	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
-	prysmv2 "github.com/prysmaticlabs/prysm/proto/prysm/v2"
-	"github.com/prysmaticlabs/prysm/shared/bytesutil"
-	"github.com/prysmaticlabs/prysm/shared/params"
-)
-
-// GenesisBeaconState returns the genesis beacon state.
-func GenesisBeaconState(ctx context.Context, deposits []*ethpb.Deposit, genesisTime uint64, eth1Data *ethpb.Eth1Data) (iface.BeaconStateAltair, error) {
-	state, err := EmptyGenesisState()
-	if err != nil {
-		return nil, err
-	}
-
-	// Process initial deposits.
-	state, err = helpers.UpdateGenesisEth1Data(state, deposits, eth1Data)
-	if err != nil {
-		return nil, err
-	}
-
-	state, err = ProcessPreGenesisDeposits(ctx, state, deposits)
-	if err != nil {
-		return nil, errors.Wrap(err, "could not process validator deposits")
-	}
-
-	return OptimizedGenesisBeaconState(genesisTime, state, state.Eth1Data())
-}
-
-// OptimizedGenesisBeaconState is used to create a state that has already processed deposits. This is to efficiently
-// create a mainnet state at chainstart.
-func OptimizedGenesisBeaconState(genesisTime uint64, preState iface.BeaconStateAltair, eth1Data *ethpb.Eth1Data) (iface.BeaconStateAltair, error) {
-	if eth1Data == nil {
-		return nil, errors.New("no eth1data provided for genesis state")
-	}
-
-	randaoMixes := make([][]byte, params.BeaconConfig().EpochsPerHistoricalVector)
-	for i := 0; i < len(randaoMixes); i++ {
-		h := make([]byte, 32)
-		copy(h, eth1Data.BlockHash)
-		randaoMixes[i] = h
-	}
-
-	zeroHash := params.BeaconConfig().ZeroHash[:]
-
-	activeIndexRoots := make([][]byte, params.BeaconConfig().EpochsPerHistoricalVector)
-	for i := 0; i < len(activeIndexRoots); i++ {
-		activeIndexRoots[i] = zeroHash
-	}
-
-	blockRoots := make([][]byte, params.BeaconConfig().SlotsPerHistoricalRoot)
-	for i := 0; i < len(blockRoots); i++ {
-		blockRoots[i] = zeroHash
-	}
-
-	stateRoots := make([][]byte, params.BeaconConfig().SlotsPerHistoricalRoot)
-	for i := 0; i < len(stateRoots); i++ {
-		stateRoots[i] = zeroHash
-	}
-
-	slashings := make([]uint64, params.BeaconConfig().EpochsPerSlashingsVector)
-
-	genesisValidatorsRoot, err := v1.ValidatorRegistryRoot(preState.Validators())
-	if err != nil {
-		return nil, errors.Wrapf(err, "could not hash tree root genesis validators %v", err)
-	}
-
-	prevEpochParticipation, err := preState.PreviousEpochParticipation()
-	if err != nil {
-		return nil, err
-	}
-	currEpochParticipation, err := preState.CurrentEpochParticipation()
-	if err != nil {
-		return nil, err
-	}
-	scores, err := preState.InactivityScores()
-	if err != nil {
-		return nil, err
-	}
-	state := &pb.BeaconStateAltair{
-		// Misc fields.
-		Slot:                  0,
-		GenesisTime:           genesisTime,
-		GenesisValidatorsRoot: genesisValidatorsRoot[:],
-
-		Fork: &pb.Fork{
-			PreviousVersion: params.BeaconConfig().GenesisForkVersion,
-			CurrentVersion:  params.BeaconConfig().GenesisForkVersion,
-			Epoch:           0,
-		},
-
-		// Validator registry fields.
-		Validators:                 preState.Validators(),
-		Balances:                   preState.Balances(),
-		PreviousEpochParticipation: prevEpochParticipation,
-		CurrentEpochParticipation:  currEpochParticipation,
-		InactivityScores:           scores,
-
-		// Randomness and committees.
-		RandaoMixes: randaoMixes,
-
-		// Finality.
-		PreviousJustifiedCheckpoint: &ethpb.Checkpoint{
-			Epoch: 0,
-			Root:  params.BeaconConfig().ZeroHash[:],
-		},
-		CurrentJustifiedCheckpoint: &ethpb.Checkpoint{
-			Epoch: 0,
-			Root:  params.BeaconConfig().ZeroHash[:],
-		},
-		JustificationBits: []byte{0},
-		FinalizedCheckpoint: &ethpb.Checkpoint{
-			Epoch: 0,
-			Root:  params.BeaconConfig().ZeroHash[:],
-		},
-
-		HistoricalRoots: [][]byte{},
-		BlockRoots:      blockRoots,
-		StateRoots:      stateRoots,
-		Slashings:       slashings,
-
-		// Eth1 data.
-		Eth1Data:         eth1Data,
-		Eth1DataVotes:    []*ethpb.Eth1Data{},
-		Eth1DepositIndex: preState.Eth1DepositIndex(),
-	}
-
-	bodyRoot, err := (&prysmv2.BeaconBlockBody{
-		RandaoReveal: make([]byte, 96),
-		Eth1Data: &ethpb.Eth1Data{
-			DepositRoot: make([]byte, 32),
-			BlockHash:   make([]byte, 32),
-		},
-		Graffiti: make([]byte, 32),
-		SyncAggregate: &prysmv2.SyncAggregate{
-			SyncCommitteeBits:      make([]byte, len(bitfield.NewBitvector512())),
-			SyncCommitteeSignature: make([]byte, 96),
-		},
-	}).HashTreeRoot()
-	if err != nil {
-		return nil, errors.Wrap(err, "could not hash tree root empty block body")
-	}
-
-	state.LatestBlockHeader = &ethpb.BeaconBlockHeader{
-		ParentRoot: zeroHash,
-		StateRoot:  zeroHash,
-		BodyRoot:   bodyRoot[:],
-	}
-
-	var pubKeys [][]byte
-	for i := uint64(0); i < params.BeaconConfig().SyncCommitteeSize; i++ {
-		pubKeys = append(pubKeys, bytesutil.PadTo([]byte{}, params.BeaconConfig().BLSPubkeyLength))
-	}
-	state.CurrentSyncCommittee = &pb.SyncCommittee{
-		Pubkeys:         pubKeys,
-		AggregatePubkey: bytesutil.PadTo([]byte{}, params.BeaconConfig().BLSPubkeyLength),
-	}
-	state.NextSyncCommittee = &pb.SyncCommittee{
-		Pubkeys:         bytesutil.Copy2dBytes(pubKeys),
-		AggregatePubkey: bytesutil.PadTo([]byte{}, params.BeaconConfig().BLSPubkeyLength),
-	}
-
-	return stateAltair.InitializeFromProto(state)
-}
-
-// EmptyGenesisState returns an empty beacon state hard fork 1 object.
-func EmptyGenesisState() (iface.BeaconStateAltair, error) {
-	state := &pb.BeaconStateAltair{
-		// Misc fields.
-		Slot: 0,
-		Fork: &pb.Fork{
-			PreviousVersion: params.BeaconConfig().GenesisForkVersion,
-			CurrentVersion:  params.BeaconConfig().GenesisForkVersion,
-			Epoch:           0,
-		},
-		// Validator registry fields.
-		Validators:       []*ethpb.Validator{},
-		Balances:         []uint64{},
-		InactivityScores: []uint64{},
-
-		JustificationBits:          []byte{0},
-		HistoricalRoots:            [][]byte{},
-		CurrentEpochParticipation:  []byte{},
-		PreviousEpochParticipation: []byte{},
-
-		// Eth1 data.
-		Eth1Data:         &ethpb.Eth1Data{},
-		Eth1DataVotes:    []*ethpb.Eth1Data{},
-		Eth1DepositIndex: 0,
-	}
-	return stateAltair.InitializeFromProto(state)
-}
