+++ beacon-chain/rpc/prysm/v1alpha1/validator/assignments_test.go
@@ -120,7 +120,7 @@ func TestGetAltairDuties_SyncCommitteeOK(t *testing.T) {
 	require.NoError(t, err)
 	eth1Data, err := testutil.DeterministicEth1Data(len(deposits))
 	require.NoError(t, err)
-	bs, err := altair.GenesisBeaconState(context.Background(), deposits, 0, eth1Data)
+	bs, err := testutil.GenesisBeaconState(context.Background(), deposits, 0, eth1Data)
 	require.NoError(t, err, "Could not setup genesis bs")
 	genesisRoot, err := genesis.Block.HashTreeRoot()
 	require.NoError(t, err, "Could not get signing root")
